refactor: move styles into common directory
