+++ src/components/header/MultilineHeaderTitle.tsx
@@ -4,7 +4,7 @@
 
 import React, { memo } from "react";
 import { GestureResponderEvent, Text, TouchableOpacity } from "react-native";
-import { headerStyles } from "../../layout/default";
+import { headerStyles } from "../../styles/layout/default";
 
 /** The props for a multiline header */
 interface IMultilineHeaderProps {
