+++ src/views/today/ClassesView.tsx
@@ -8,8 +8,8 @@ import React from "react";
 import { ScrollView } from "react-native";
 import AdvisoryComponent from "../../components/blocks/AdvisoryComponent";
 import BlockComponent from "../../components/blocks/BlockComponent";
-import { settingsViewStyles } from "../../layout/default";
 import { PreparedClasses } from "../../stores/preparedClassesStore";
+import { settingsViewStyles } from "../../styles/layout/default";
 import { Block } from "../../util/blocks/block";
 import { ICalendarSchoolDay, SchoolDay } from "../../util/calendar/types";
 import { IAdvisory } from "../../util/class/classes";
